ebisp/: Use ATOM_LAMBDA instead of bare "lambda" forms
  * First off, this gives us faster lambda_p().  No more parsing
list structure to determine if we have a lambda each time we call
a function, it becomes a simple type-test.

  * Second, this now gives us a single mechanism for creating
lambdas, making it easier to modify the structure of a lambda in
the future (such as for storing a closure environment).

  * Third, this is a semantic change to the language: We can no
longer construct a valid lambda using calls to (list), (quote), or
(quasiquote); they must always be created by (defun), (lambda), or
(λ).