ebisp/scope: Re-use Scope CONSes where possible
  * First, by mutating existing CONSes ("value cells") when we set
a value in a scope with an existing binding, we prevent runaway
allocation of un-collectable garbage.  If we simply shadow the
existing binding then it can't be reclaimed until the new binding
is discarded, and at global scope that means that it can't be
reclaimed until the global scope is destroyed.

  * Second, by preserving the "spine" of the environment list, we
can "close over" that environment by taking a reference to the
"spine", and can view all subsequent changes (outside of a new
binding being added by push_scope_frame()).

  * There seems to be an impossible case in set_scope_value_impl():
scope.expr is initialized to a CONS and only possibly changed to
something else by pop_scope_frame(), which is only used in a pair
with push_scope_frame(), thus having no net effect on the type.
So how can it ever not be a CONS?