ebisp/gc: Prevent stack blow-out in gc_traverse_expr() on cyclic structure
  * If we manage to construct a cyclic list structure and then
attempt to run a GC while any part of the cycle is still "live",
gc_traverse_expr() was attempting to traverse the entire cycle to
the end before marking any given cell in the cycle as having been
visited.

  * Mark cells as being visited before attempting to visit the
values to which they refer, causing gc_traverse_expr() to return
without further processing when it encounters a cycle.

  * This might-or-might-not be issue #578.  There's not enough
context in the issue description for me to be certain.