ebisp/interpreter: Use closed-over environment as basis for lambda calls
  * Instead of extending our caller Scope with a frame containing
the arguments for the new function, create a new Scope from the
environment stored in the lambda and extend that with the new
frame.

  * This is a MAJOR semantic change to the language.  Because the
caller environment is not passed to the callee, we end up with
lexical (instead of dynamic) scope (issue #338).  Because we use
the environment stored in the lambda, we end up with closures
(issue #620).  And because the environment mutates its value cells,
the closures are mutable as well.

  * ((lambda (x) (defun foo () (set x (+ x 1)))) 0) now creates a
global function called "foo" that returns successive integers from
1 each time it's called.  And doesn't leak unreclaimable memory
each time it's called.