(#355) Generify deleting words/characters with memmove
Calling `delete_char()` in a loop is wasteful, as it calls `memmove()`
internally over and over. This is especially inefficient for large
words.

Adds a generic `kill_region_and_move_cursor()` method, which destroys
characters between [start, end] (inclusive), taking into account when
to move with `memmove()` and when to simply kill to the end of the
line. The cursor is then moved to `start`. Now, deleting a single
character is simply deleting a region of size 1.